+++ include/world/objects/RenderIntentElement.h
@@ -34,6 +34,8 @@ class RenderIntentElement : public Element {
                setRaytracerMaxRecursionDepth)
   Q_PROPERTY(int pathTracerRussianRouletteDepth READ pathTracerRussianRouletteDepth WRITE
                setPathTracerRussianRouletteDepth)
+  Q_PROPERTY(int pathTracerDirectLightSamples READ pathTracerDirectLightSamples WRITE
+               setPathTracerDirectLightSamples)
   Q_PROPERTY(QString raytracerViewPlane READ raytracerViewPlane WRITE setRaytracerViewPlane)
   Q_PROPERTY(int raytracerThreads READ raytracerThreads WRITE setRaytracerThreads)
   Q_PROPERTY(int raytracerQueueSize READ raytracerQueueSize WRITE setRaytracerQueueSize)
@@ -139,6 +141,9 @@ public:
 
   int pathTracerRussianRouletteDepth() const;
   void setPathTracerRussianRouletteDepth(int depth);
+
+  int pathTracerDirectLightSamples() const;
+  void setPathTracerDirectLightSamples(int samples);
 
   QString raytracerViewPlane() const;
   void setRaytracerViewPlane(const QString& viewPlane);